Added clarifying comments for jibe controller methods and renamed class
