+++ python/sail_controller.py
@@ -21,3 +21,5 @@ class SailController:
             * abs(bounded_angle)
             + sailbot_constants.SAIL_CONTROLLER_MAX_SAIL_ANGLE
         )
+
+
