+++ test/test_JibeController.py
@@ -5,7 +5,7 @@ from __future__ import division
 from unittest.case import TestCase
 
 import local_imports
-#import rostest
+import rostest
 import unittest
 import math
 import sailbot_constants
@@ -41,6 +41,5 @@ class Test_JibeController(unittest.TestCase):
     def test_getJibeError_wrap(self):
         self.assertAlmostEqual(JibeOnlySailController.get_jibe_controller_error(math.pi /2 + 2 * math.pi, - math.pi / 2 + 2 * math.pi, -1), -math.pi)
 if __name__ == "__main__":
-    #rostest.rosrun("boat_controller", "Test_SailController",
-     #              Test_SailController)
-    unittest.main()
+    rostest.rosrun("jibe_only_rudder_controller", "Test_JibeController",
+                    Test_JibeController)
